Split boxmon_expression to separate file, fix build
